WIP gemm local
